Add the instructor side. For now default is instructor until I figure out cookies
