+++ Pages/Home/Create.cshtml
@@ -67,7 +67,7 @@
 </div>
 
 <div>
-    <a asp-page="Index">Back to List</a>
+    <a asp-page="InstructorDashboard">Back to List</a>
 </div>
 
 @section Scripts {
