+++ Pages/Home/Edit.cshtml
@@ -62,7 +62,7 @@
 </div>
 
 <div>
-    <a asp-page="./Index">Back to List</a>
+    <a asp-page="./InstructorDashboard">Back to List</a>
 </div>
 
 @section Scripts {
