+++ Pages/Home/InstructorDetails.cshtml
@@ -1,13 +1,17 @@
 ﻿@page
-@model Assignment1v3.Pages.Home.DeleteModel
+@model Assignment1v3.Pages.Home.InstructorDetailsModel
 
 @{
-    ViewData["Title"] = "Delete";
+    ViewData["Title"] = "Details";
 }
 
-<h1>Delete</h1>
+<h1>Details</h1>
+<style>
+    .red-link {
+        color: red;
+    }
+</style>
 
-<h3>Are you sure you want to delete this?</h3>
 <div>
     <h4>Course</h4>
     <hr />
@@ -61,10 +65,9 @@
             @Html.DisplayFor(model => model.Course.Description)
         </dd>
     </dl>
-    
-    <form method="post">
-        <input type="hidden" asp-for="Course.Id" />
-        <input type="submit" value="Delete" class="btn btn-danger" /> |
-        <a asp-page="./InstructorDashboard">Back to List</a>
-    </form>
+</div>
+<div>
+    <a asp-page="./Edit" asp-route-id="@Model.Course?.Id">Edit</a> |
+    <a class="red-link" asp-page="./Delete" asp-route-id="@Model.Course?.Id">Delete</a> |
+    <a asp-page="./InstructorDashboard">Back to List</a>
 </div>
