+++ Pages/Home/StudentDashboard.cshtml
@@ -1,12 +1,12 @@
 ﻿@page
-@model Assignment1v3.Pages.Home.IndexModel
+@model Assignment1v3.Pages.Home.StudentDashboardModel
 
 @{
-    ViewData["Title"] = "Index";
+    ViewData["Title"] = "Student Dashboard";
 }
 
 <div class="container">
-    <h1>Dashboard</h1>
+    <h1>Student Dashboard</h1>
 
     <div class="row">
         <div class="col-md-9">
@@ -24,6 +24,9 @@
                             <p class="card-text">@course.CourseName</p>
                             <p class="card-text">@course.Location</p>
                             <p class="card-text">@course.StartTime</p>
+                            <a class="btn btn-primary" asp-page="./StudentDetails" asp-route-id="@course.Id">Details</a>
+                            <!--Need the drop stuff-->
+                            <a class="btn btn-danger" asp-page="./StudentDetails" asp-route-id="@course.Id">Drop</a>
                         </div>
                     </div>
                 </div>
@@ -37,6 +40,7 @@
                     <!-- Add To-Do list here -->
                 </div>
             </div>
+        </div>
         </div>
     </div>
 </div>
