+++ Pages/Home/StudentDashboard.cshtml.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.EntityFrameworkCore;
+using Assignment1v3.Data;
+using Assignment1v3.Models;
+using Microsoft.AspNetCore.Authorization;
+
+namespace Assignment1v3.Pages.Home
+{
+    //[Authorize(Policy = "MustBeStudent")] 
+    public class StudentDashboardModel : PageModel
+    {
+        private readonly Assignment1v3Context _context;
+
+        public StudentDashboardModel(Assignment1v3Context context)
+        {
+            _context = context;
+        }
+
+        public IList<Course> Course { get; set; }
+
+        public async Task OnGetAsync()
+        {
+            Course = await _context.Course.ToListAsync();
+        }
+    }
+}
