+++ Pages/Home/StudentDetails.cshtml.cs
@@ -10,17 +10,16 @@ using Assignment1v3.Models;
 
 namespace Assignment1v3.Pages.Home
 {
-    public class DeleteModel : PageModel
+    public class StudentDetailsModel : PageModel
     {
         private readonly Assignment1v3.Data.Assignment1v3Context _context;
 
-        public DeleteModel(Assignment1v3.Data.Assignment1v3Context context)
+        public StudentDetailsModel(Assignment1v3.Data.Assignment1v3Context context)
         {
             _context = context;
         }
 
-        [BindProperty]
-      public Course Course { get; set; } = default!;
+      public Course Course { get; set; } = default!; 
 
         public async Task<IActionResult> OnGetAsync(int? id)
         {
@@ -30,7 +29,6 @@ namespace Assignment1v3.Pages.Home
             }
 
             var course = await _context.Course.FirstOrDefaultAsync(m => m.Id == id);
-
             if (course == null)
             {
                 return NotFound();
@@ -40,24 +38,6 @@ namespace Assignment1v3.Pages.Home
                 Course = course;
             }
             return Page();
-        }
-
-        public async Task<IActionResult> OnPostAsync(int? id)
-        {
-            if (id == null || _context.Course == null)
-            {
-                return NotFound();
-            }
-            var course = await _context.Course.FindAsync(id);
-
-            if (course != null)
-            {
-                Course = course;
-                _context.Course.Remove(Course);
-                await _context.SaveChangesAsync();
-            }
-
-            return RedirectToPage("./InstructorDashboard");
         }
     }
 }
