+++ Pages/Shared/_Layout.cshtml
@@ -28,7 +28,7 @@
                     <ul class="navbar-nav flex-grow-1 mr-auto">
                         
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-page="/Home/Index">Dashboard</a>
+                            <a class="nav-link text-dark" asp-area="" asp-page="/Home/InstructorDashboard">Dashboard</a>
                         </li>
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-page="/Logins/Create">Sign Up</a>
